# tp: speed up JSON tokenization/parsing by ~9x