--- a/src/trace_processor/importers/common/BUILD.gn
+++ b/src/trace_processor/importers/common/BUILD.gn
@@ -107,7 +107,10 @@ source_set("common") {
 
 source_set("trace_parser_hdr") {
   sources = [ "trace_parser.h" ]
-  deps = [ "../../../../gn:default_deps" ]
+  deps = [
+    ":parser_types",
+    "../../../../gn:default_deps",
+  ]
 }
 
 source_set("parser_types") {