--- a/src/trace_processor/importers/common/parser_types.h
+++ b/src/trace_processor/importers/common/parser_types.h
@@ -20,13 +20,16 @@
 #include <array>
 #include <cstdint>
 #include <functional>
+#include <limits>
+#include <memory>
 #include <optional>
-#include <string>
 #include <utility>
 #include <variant>
+#include <vector>
 
 #include "perfetto/trace_processor/ref_counted.h"
 #include "perfetto/trace_processor/trace_blob_view.h"
+#include "src/trace_processor/containers/interval_tree.h"
 #include "src/trace_processor/containers/string_pool.h"
 #include "src/trace_processor/importers/proto/packet_sequence_state_generation.h"
 
@@ -57,16 +60,72 @@ struct alignas(8) InlineSchedWaking {
 static_assert(sizeof(InlineSchedWaking) == 16);
 
 struct alignas(8) JsonEvent {
-  struct Begin {};
-  struct End {};
-  struct Scoped {
-    int64_t dur;
-  };
-  struct Other {};
-  using Type = std::variant<Begin, End, Scoped, Other>;
+  int64_t dur = std::numeric_limits<int64_t>::max();
 
-  std::string value;
-  Type type;
+  uint32_t pid = 0;
+  uint32_t tid = 0;
+
+  StringPool::Id name = StringPool::Id::Null();
+  StringPool::Id cat = StringPool::Id::Null();
+
+  union IdStrOrUint64 {
+    StringPool::Id id_str;
+    uint64_t id_uint64;
+  };
+  IdStrOrUint64 id;
+  IdStrOrUint64 bind_id;
+
+  int64_t tts = std::numeric_limits<int64_t>::max();
+  int64_t tdur = std::numeric_limits<int64_t>::max();
+  int64_t async_cookie = std::numeric_limits<int64_t>::max();
+
+  std::unique_ptr<char[]> args;
+
+  char phase = '\0';
+
+  uint64_t flow_in : 1;
+  uint64_t flow_out : 1;
+
+  uint64_t pid_is_string_id : 1;
+  uint64_t tid_is_string_id : 1;
+
+  uint64_t bind_enclosing_slice : 1;
+
+  enum class IdType : uint64_t {
+    kNone = 0,
+    kString = 1,
+    kUint64 = 2,
+  };
+  IdType id_type : 2;
+  IdType bind_id_type : 2;
+
+  enum class Scope : uint64_t {
+    kNone = 0,
+    kGlobal = 1,
+    kProcess = 2,
+    kThread = 3,
+  };
+  Scope scope : 2;
+
+  enum class AsyncCookieType : uint64_t {
+    kId,
+    kId2Local,
+    kId2Global,
+  };
+  AsyncCookieType async_cookie_type : 2;
+
+  uint64_t args_size : 43;
+
+  JsonEvent()
+      : flow_in(false),
+        flow_out(false),
+        pid_is_string_id(false),
+        tid_is_string_id(false),
+        bind_enclosing_slice(false),
+        id_type(IdType::kNone),
+        bind_id_type(IdType::kNone),
+        scope(Scope::kNone),
+        args_size(0) {}
 };
 static_assert(sizeof(JsonEvent) % 8 == 0);