--- a/src/trace_processor/importers/json/json_utils.cc
+++ b/src/trace_processor/importers/json/json_utils.cc
@@ -17,127 +17,34 @@
 #include "src/trace_processor/importers/json/json_utils.h"
 
 #include "perfetto/base/build_config.h"
+#include "perfetto/base/compiler.h"
+#include "perfetto/base/logging.h"
+#include "perfetto/ext/base/small_vector.h"
+#include "perfetto/ext/base/string_view.h"
+#include "src/trace_processor/containers/string_pool.h"
+#include "src/trace_processor/importers/common/args_tracker.h"
+#include "src/trace_processor/importers/json/json_parser.h"
+#include "src/trace_processor/storage/trace_storage.h"
+#include "src/trace_processor/types/variadic.h"
 
-#include <limits>
+#include <cstddef>
+#include <cstdint>
+#include <cstdlib>
+#include <memory>
+#include <optional>
+#include <string>
+#include <string_view>
+#include <utility>
+#include <vector>
 
 #if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
 #include <json/reader.h>
-#include "perfetto/ext/base/string_utils.h"
+#include <json/value.h>
 #endif
 
-namespace perfetto {
-namespace trace_processor {
-namespace json {
-
-std::optional<int64_t> CoerceToTs(const Json::Value& value) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
-#if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
-  switch (static_cast<size_t>(value.type())) {
-    case Json::realValue:
-      return static_cast<int64_t>(value.asDouble() * 1000.0);
-    case Json::uintValue:
-    case Json::intValue:
-      return value.asInt64() * 1000;
-    case Json::stringValue:
-      return CoerceToTs(value.asString());
-    default:
-      return std::nullopt;
-  }
-#else
-  perfetto::base::ignore_result(value);
-  return std::nullopt;
-#endif
-}
-
-std::optional<int64_t> CoerceToTs(const std::string& s) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
-#if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
-  // 's' is formatted as a JSON Number, in microseconds
-  // goal: reformat 's' to be as an int, in nanoseconds
-  std::string s_as_ns = s;
-
-  // detect and remove scientific notation's exponents
-  int32_t exp_shift = 0;
-  if (size_t exp_start = s.find_first_of("eE");
-      exp_start != std::string::npos) {
-    const std::string exp_s = s.substr(exp_start + 1, s.size());
-    const std::optional<int32_t> exp = base::StringToInt32(exp_s);
-    if (!exp.has_value()) {
-      return std::nullopt;
-    }
-    s_as_ns.erase(exp_start);
-    exp_shift = *exp;
-  }
-
-  // detect and remove decimal separator
-  size_t int_size = s_as_ns.size();
-  if (size_t frac_start = s.find('.'); frac_start != std::string::npos) {
-    s_as_ns.erase(frac_start, 1);
-    int_size = frac_start;
-  }
-
-  // expand or shrink to the new size
-  constexpr int us_to_ns_shift = 3;
-  const size_t s_as_ns_size = size_t(
-      std::max<ptrdiff_t>(1, ptrdiff_t(int_size) + exp_shift + us_to_ns_shift));
-  s_as_ns.resize(s_as_ns_size, '0');  // pads or truncates
-
-  return base::StringToInt64(s_as_ns);
-#else
-  perfetto::base::ignore_result(s);
-  return std::nullopt;
-#endif
-}
-
-std::optional<int64_t> CoerceToInt64(const Json::Value& value) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
-#if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
-  switch (static_cast<size_t>(value.type())) {
-    case Json::realValue:
-    case Json::uintValue:
-      return static_cast<int64_t>(value.asUInt64());
-    case Json::intValue:
-      return value.asInt64();
-    case Json::stringValue: {
-      std::string s = value.asString();
-      char* end;
-      int64_t n = strtoll(s.c_str(), &end, 10);
-      if (end != s.data() + s.size())
-        return std::nullopt;
-      return n;
-    }
-    default:
-      return std::nullopt;
-  }
-#else
-  perfetto::base::ignore_result(value);
-  return std::nullopt;
-#endif
-}
-
-std::optional<uint32_t> CoerceToUint32(const Json::Value& value) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
-#if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
-  std::optional<int64_t> result = CoerceToInt64(value);
-  if (!result.has_value())
-    return std::nullopt;
-  int64_t n = result.value();
-  if (n < 0 || n > std::numeric_limits<uint32_t>::max())
-    return std::nullopt;
-  return static_cast<uint32_t>(n);
-#else
-  perfetto::base::ignore_result(value);
-  return std::nullopt;
-#endif
-}
+namespace perfetto::trace_processor::json {
 
 std::optional<Json::Value> ParseJsonString(base::StringView raw_string) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
 #if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
   Json::CharReaderBuilder b;
   auto reader = std::unique_ptr<Json::CharReader>(b.newCharReader());
@@ -153,87 +60,112 @@ std::optional<Json::Value> ParseJsonString(base::StringView raw_string) {
 #endif
 }
 
-bool AddJsonValueToArgs(const Json::Value& value,
-                        base::StringView flat_key,
-                        base::StringView key,
+bool AddJsonValueToArgs(Iterator& it,
+                        const char* start,
+                        const char* end,
+                        std::string_view flat_key,
+                        std::string_view key,
                         TraceStorage* storage,
                         ArgsTracker::BoundInserter* inserter) {
-  PERFETTO_DCHECK(IsJsonSupported());
-
-#if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
-  if (value.isObject()) {
-    auto it = value.begin();
+  it.Reset(start, end);
+  if (!it.ParseStart()) {
+    return false;
+  }
+  struct Frame {
+    uint32_t flat_key_size = 0;
+    uint32_t key_size = 0;
     bool inserted = false;
-    for (; it != value.end(); ++it) {
-      std::string child_name = it.name();
-      std::string child_flat_key = flat_key.ToStdString() + "." + child_name;
-      std::string child_key = key.ToStdString() + "." + child_name;
-      inserted |=
-          AddJsonValueToArgs(*it, base::StringView(child_flat_key),
-                             base::StringView(child_key), storage, inserter);
+  };
+  base::SmallVector<Frame, 4> stack;
+  std::string flat_key_str(flat_key);
+  std::string key_str(key);
+
+  stack.emplace_back(Frame{static_cast<uint32_t>(flat_key_str.size()),
+                           static_cast<uint32_t>(key_str.size()), false});
+  for (;;) {
+    json::Iterator::ParseType parse_type = it.parse_stack().back();
+    auto& frame = stack.back();
+    switch (it.ParseAndRecurse()) {
+      case json::ReturnCode::kOk:
+        break;
+      case json::ReturnCode::kEndOfScope: {
+        PERFETTO_DCHECK(!stack.empty());
+        bool inserted = frame.inserted;
+        stack.pop_back();
+        if (stack.empty()) {
+          return inserted;
+        }
+        auto& new_frame = stack.back();
+        new_frame.inserted |= inserted;
+        continue;
+      }
+      case json::ReturnCode::kIncompleteInput:
+      case json::ReturnCode::kError:
+        return false;
     }
-    return inserted;
-  }
 
-  if (value.isArray()) {
-    auto it = value.begin();
-    bool inserted_any = false;
-    std::string array_key = key.ToStdString();
-    StringId array_key_id = storage->InternString(key);
-    for (; it != value.end(); ++it) {
-      size_t array_index = inserter->GetNextArrayEntryIndex(array_key_id);
-      std::string child_key =
-          array_key + "[" + std::to_string(array_index) + "]";
-      bool inserted = AddJsonValueToArgs(
-          *it, flat_key, base::StringView(child_key), storage, inserter);
-      if (inserted)
-        inserter->IncrementArrayEntryIndex(array_key_id);
-      inserted_any |= inserted;
+    // Skip null values.
+    const auto& value = it.value();
+    if (value.index() == base::variant_index<json::JsonValue, json::Null>()) {
+      continue;
     }
-    return inserted_any;
-  }
 
-  // Leaf value.
-  auto flat_key_id = storage->InternString(flat_key);
-  auto key_id = storage->InternString(key);
+    // Leaf value.
+    flat_key_str.resize(frame.flat_key_size);
+    key_str.resize(frame.key_size);
 
-  switch (value.type()) {
-    case Json::ValueType::nullValue:
-      break;
-    case Json::ValueType::intValue:
-      inserter->AddArg(flat_key_id, key_id, Variadic::Integer(value.asInt64()));
-      return true;
-    case Json::ValueType::uintValue:
-      inserter->AddArg(flat_key_id, key_id,
-                       Variadic::UnsignedInteger(value.asUInt64()));
-      return true;
-    case Json::ValueType::realValue:
-      inserter->AddArg(flat_key_id, key_id, Variadic::Real(value.asDouble()));
-      return true;
-    case Json::ValueType::stringValue:
-      inserter->AddArg(flat_key_id, key_id,
-                       Variadic::String(storage->InternString(
-                           base::StringView(value.asString()))));
-      return true;
-    case Json::ValueType::booleanValue:
-      inserter->AddArg(flat_key_id, key_id, Variadic::Boolean(value.asBool()));
-      return true;
-    case Json::ValueType::objectValue:
-    case Json::ValueType::arrayValue:
-      PERFETTO_FATAL("Non-leaf types handled above");
-      break;
+    if (parse_type == json::Iterator::ParseType::kArray) {
+      StringPool::Id key_id = storage->InternString(key_str);
+      size_t array_index = inserter->GetNextArrayEntryIndex(key_id);
+      key_str += "[" + std::to_string(array_index) + "]";
+      inserter->IncrementArrayEntryIndex(key_id);
+    } else if (parse_type == json::Iterator::ParseType::kObject) {
+      key_str += ".";
+      key_str += it.key();
+      flat_key_str += ".";
+      flat_key_str += it.key();
+    } else {
+      PERFETTO_FATAL("Unexpected parse type %d", static_cast<int>(parse_type));
+    }
+
+    if (value.index() == base::variant_index<json::JsonValue, json::Object>() ||
+        value.index() == base::variant_index<json::JsonValue, json::Array>()) {
+      stack.emplace_back(Frame{static_cast<uint32_t>(flat_key_str.size()),
+                               static_cast<uint32_t>(key_str.size()), false});
+      continue;
+    }
+
+    // Only for leaf values we actually insert into the args table.
+    frame.inserted = true;
+
+    StringPool::Id flat_key_id = storage->InternString(flat_key_str);
+    StringPool::Id key_id = storage->InternString(key_str);
+    switch (it.value().index()) {
+      case base::variant_index<json::JsonValue, double>():
+        inserter->AddArg(
+            flat_key_id, key_id,
+            Variadic::Real(base::unchecked_get<double>(it.value())));
+        break;
+      case base::variant_index<json::JsonValue, int64_t>():
+        inserter->AddArg(
+            flat_key_id, key_id,
+            Variadic::Integer(base::unchecked_get<int64_t>(it.value())));
+        break;
+      case base::variant_index<json::JsonValue, std::string_view>():
+        inserter->AddArg(
+            flat_key_id, key_id,
+            Variadic::String(storage->InternString(
+                base::unchecked_get<std::string_view>(it.value()))));
+        break;
+      case base::variant_index<json::JsonValue, bool>():
+        inserter->AddArg(
+            flat_key_id, key_id,
+            Variadic::Boolean(base::unchecked_get<bool>(it.value())));
+        break;
+      default:
+        PERFETTO_FATAL("Unreachable");
+    }
   }
-  return false;
-#else
-  perfetto::base::ignore_result(value);
-  perfetto::base::ignore_result(flat_key);
-  perfetto::base::ignore_result(key);
-  perfetto::base::ignore_result(storage);
-  perfetto::base::ignore_result(inserter);
-  return false;
-#endif
 }
 
-}  // namespace json
-}  // namespace trace_processor
-}  // namespace perfetto
+}  // namespace perfetto::trace_processor::json