--- a/src/trace_processor/importers/json/json_utils.h
+++ b/src/trace_processor/importers/json/json_utils.h
@@ -17,13 +17,13 @@
 #ifndef SRC_TRACE_PROCESSOR_IMPORTERS_JSON_JSON_UTILS_H_
 #define SRC_TRACE_PROCESSOR_IMPORTERS_JSON_JSON_UTILS_H_
 
-#include <cstdint>
 #include <optional>
-#include <string>
+#include <string_view>
 
 #include "perfetto/base/build_config.h"
 #include "perfetto/ext/base/string_view.h"
 #include "src/trace_processor/importers/common/args_tracker.h"
+#include "src/trace_processor/importers/json/json_parser.h"
 #include "src/trace_processor/storage/trace_storage.h"
 
 #if PERFETTO_BUILDFLAG(PERFETTO_TP_JSON)
@@ -46,24 +46,22 @@ constexpr bool IsJsonSupported() {
 #endif
 }
 
-std::optional<int64_t> CoerceToTs(const Json::Value& value);
-std::optional<int64_t> CoerceToTs(const std::string& value);
-std::optional<int64_t> CoerceToInt64(const Json::Value& value);
-std::optional<uint32_t> CoerceToUint32(const Json::Value& value);
-
 // Parses the given JSON string into a JSON::Value object.
 // This function should only be called if |IsJsonSupported()| returns true.
 std::optional<Json::Value> ParseJsonString(base::StringView raw_string);
 
-// Flattens the given Json::Value and adds each leaf node to the bound args
-// inserter. Note:
+// Flattens the given serialized json value (with bounds `start` and `end`)
+// using `it` and adds each leaf node to the bound args inserter.
+//
+// Note:
 //  * |flat_key| and |key| should be non-empty and will be used to prefix the
 //    keys of all leaf nodes in the JSON.
 //  * |storage| is used to intern all strings (e.g. keys and values).
-//  * This function should only be called if |IsJsonSupported()| returns true.
-bool AddJsonValueToArgs(const Json::Value& value,
-                        base::StringView flat_key,
-                        base::StringView key,
+bool AddJsonValueToArgs(Iterator& it,
+                        const char* start,
+                        const char* end,
+                        std::string_view flat_key,
+                        std::string_view key,
                         TraceStorage* storage,
                         ArgsTracker::BoundInserter* inserter);