… src/trace_processor/importers/json/json_parser.h
… src/trace_processor/importers/json/json_trace_tokenizer.cc
… src/trace_processor/importers/json/json_utils_unittest.cc
… src/trace_processor/importers/proto/args_parser.cc
… src/trace_processor/sorter/trace_sorter.h
… src/trace_processor/util/debug_annotation_parser.cc
… test/trace_processor/diff_tests/parser/json/tests.py