This CL moves away from using jsoncpp as much as possible in trace processor for performance reasons. Turns out jsoncpp really sucks very badly and with a few hundred lines of code, we can destroy it in benchmarks.

Before:
```
[898.916] processor_shell.cc:1878 Trace loaded: 453.80 MB in 22.06s (20.6 MB/s)
```

After:
```
[928.441] processor_shell.cc:1878 Trace loaded: 453.80 MB in 2.56s (177.1 MB/s)
```

There should be little-no functional change as a result of this CL. I've added a bunch of tests which pass both before and after to try and ensure there is no behavioural change. It's possible that there are some bugs of course but given the huge perf wins here, I think it's worth landing.